requirements.txt
Pillow
mutagen
pylint
pytest
python-dotenv
